pyproject.toml
[project]
name = "custom_opengl_experiments"
version = "0.0.1"
authors = [
  { name="Example Author", email="author@example.com" },
]
description = "A small example package"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

dependencies = [
  "pyopengl",
  "PyQt6",
  "numpy",
  "imageio",
  "Pillow",
  "matplotlib"
]

[project.urls]
Homepage = "https://github.com/pypa/sampleproject"
Issues = "https://github.com/pypa/sampleproject/issues"

[project.scripts]
customgl = "src:shadow_mapping"

[tool.setuptools.package-data]
customgl = ["textures/**/*.png", "textures/**/*.jpg", "drawing/shaders/*.vert", "drawing/shaders/*.geom", "drawing/shaders/*.frag"]